pyproject.toml
[tool.poetry]
name = "tt2-work"
version = "1.0.0"
description = "TT2 ML Work"
authors = ["joyanedel <joyanede@alumnos.inf.utfsm.cl>"]
readme = "README.md"
packages = [{include = "base_code"}, ]

[tool.poetry.dependencies]
python = "~3.9"
numpy = "1.25.2"
pandas = "2.1.0"
torch = "2.1.0"
ipykernel = "^6.25.2"
ruff = "^0.1.6"
scikit-learn = "^1.3.0"
avalanche-lib = "^0.4.0"
seaborn = "^0.13.0"
torch-summary = "^1.4.5"

[tool.poetry.group.dev.dependencies]
flake8 = "6.1.0"
flake8-bugbear = "23.7.10"
black = "23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[flake8]
max-line-length = 100
max-complexity = 10

[tool.ruff]
line-length = 100
ignore = ["E501"]


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"
